+++ miniconf/tests/embedded/src/main.rs
@@ -9,9 +9,7 @@ use cortex_m_rt::entry;
 use cortex_m_semihosting::{debug, hprintln};
 
 use crosstrait::{register, Cast};
-use miniconf::{
-    self, IntoKeys, JsonCoreSlash, JsonPath, Node, Packed, Path, Tree, TreeAny, TreeKey,
-};
+use miniconf::{self, json, IntoKeys, JsonPath, Node, Packed, Path, Tree, TreeAny, TreeKey};
 
 use core::ops::{AddAssign, SubAssign};
 register! { i32 => dyn AddAssign<i32> }
@@ -50,14 +48,14 @@ fn main() -> ! {
     let mut s = Settings::default();
 
     let path = Path::<_, '/'>::from("/i/1/val");
-    s.set_json(&path, b"3").unwrap();
+    json::set_by_key(&mut s, &path, b"3").unwrap();
 
     let (packed, node) = Settings::transcode::<Packed, _>(&path).unwrap();
     assert_eq!(packed.into_lsb().get(), 0b1_01_01_0);
     assert_eq!(node, Node::leaf(3));
 
     let mut buf = [0; 10];
-    let len = s.get_json_by_key(packed, &mut buf).unwrap();
+    let len = json::get_by_key(&s, packed, &mut buf).unwrap();
     assert_eq!(&buf[..len], b"3");
 
     let key = JsonPath::from(".i[1].val");