+++ CHANGELOG.md
@@ -6,11 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.14.2](https://github.com/quartiq/miniconf/compare/v0.14.1...v0.14.2) - 2024-09-30
+## [Unreleased](https://github.com/quartiq/miniconf/compare/v0.14.1...HEAD) - DATE
 
 ### Changed
 
 * `Traversal` and `Error` are now `exhaustive`.
+
+### Removed
+
+* The `JsonCoreSlash`/`JsonCoreSlashOwned` and `Postcard`/`PostcardOwned` traits have
+  been removed in favor of simple functions in the `json::` and `postcard::` modules.
 
 ## [0.14.1](https://github.com/quartiq/miniconf/compare/v0.13.0...v0.14.1) - 2024-09-30